refactor: 프로필페이지 컴포넌트 폴더 구조 변경 및 index.ts 추가
