+++ src/components/pages/profiles/index.ts
@@ -1,0 +1,1 @@
+export { default as ProfilePage } from './ProfilePage';
